% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_dd_tapr_body}
\alias{build_dd_tapr_body}
\title{Build POST body for dd_tapr request}
\usage{
build_dd_tapr_body(end_year, prgopt, level, tapr, sumlev, dsname, keys)
}
\arguments{
\item{end_year}{School year end}

\item{prgopt}{SAS program path}

\item{level}{"Campus" or "District"}

\item{tapr}{"all_c" or "all_d"}

\item{sumlev}{"C" or "D"}

\item{dsname}{"STUD" or "REF"}

\item{keys}{Character vector of key codes}
}
\value{
URL-encoded body string
}
\description{
Build POST body for dd_tapr request
}
\keyword{internal}
