% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_staar.R
\name{cache_exists_staar}
\alias{cache_exists_staar}
\title{Check if STAAR cache exists}
\usage{
cache_exists_staar(year, level, grade, subject, cache_type)
}
\arguments{
\item{year}{Assessment year}

\item{level}{"district" or "campus"}

\item{grade}{Grade level}

\item{subject}{"english" or "spanish"}

\item{cache_type}{"tidy" or "processed"}
}
\value{
Logical indicating if cache exists
}
\description{
Check if STAAR cache exists
}
\keyword{internal}
