% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_aeis_file}
\alias{download_aeis_file}
\title{Download a single AEIS data file}
\usage{
download_aeis_file(end_year, dsname, keys, id_param)
}
\arguments{
\item{end_year}{School year end}

\item{dsname}{Dataset name (cstud, dstud, cref, dref)}

\item{keys}{Vector of key names to request}

\item{id_param}{ID parameter string (e.g., "dist0=999999")}
}
\value{
Data frame
}
\description{
Download a single AEIS data file
}
\keyword{internal}
