% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\name{download_grad_file}
\alias{download_grad_file}
\title{Download a single graduation rate Excel file}
\usage{
download_grad_file(url, class_year, level)
}
\arguments{
\item{url}{URL to download}

\item{class_year}{Class year for error messages}

\item{level}{Data level ("campus", "district", or "county")}
}
\value{
Data frame with graduation rate data
}
\description{
Download a single graduation rate Excel file
}
\keyword{internal}
