% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from TEA}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
List with campus and district data frames
}
\description{
Downloads campus and district enrollment data from TEA's reporting systems.
Uses TAPR for 2013+, AEIS SAS broker for 2003-2012, and AEIS CGI for 1997-2002.
}
\keyword{internal}
