% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_staar.R
\name{get_staar_cache_path}
\alias{get_staar_cache_path}
\title{Get STAAR cache path}
\usage{
get_staar_cache_path(year, level, grade, subject, cache_type)
}
\arguments{
\item{year}{Assessment year}

\item{level}{"district" or "campus"}

\item{grade}{Grade level}

\item{subject}{"english" or "spanish"}

\item{cache_type}{"tidy" or "processed"}
}
\value{
File path for cache
}
\description{
Generates cache file path for STAAR data.
}
\keyword{internal}
