% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_staar.R
\name{get_staar_grades}
\alias{get_staar_grades}
\title{Get STAAR grade levels}
\usage{
get_staar_grades()
}
\value{
Integer vector of available grades (3-8)
}
\description{
Returns available grade levels for STAAR data.
}
