% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{parse_aeis_layout}
\alias{parse_aeis_layout}
\title{Parse AEIS layout file to extract column names}
\usage{
parse_aeis_layout(layout_text)
}
\arguments{
\item{layout_text}{Text content of layout file}
}
\value{
Character vector of column names
}
\description{
Parse AEIS layout file to extract column names
}
\keyword{internal}
