% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{pivot_adhoc_ethnicity}
\alias{pivot_adhoc_ethnicity}
\title{Pivot ad-hoc ethnicity data to wide format matching TAPR columns}
\usage{
pivot_adhoc_ethnicity(df, sumlev)
}
\arguments{
\item{df}{Long-format data frame from ad-hoc ethnicity download}

\item{sumlev}{"C" for campus, "D" for district}
}
\value{
Wide data frame with TAPR-compatible column names
}
\description{
Pivot ad-hoc ethnicity data to wide format matching TAPR columns
}
\keyword{internal}
