% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_campus_enr}
\alias{process_campus_enr}
\title{Process campus-level enrollment data}
\usage{
process_campus_enr(df, end_year)
}
\arguments{
\item{df}{Raw campus data frame}

\item{end_year}{School year end}
}
\value{
Processed campus data frame
}
\description{
Process campus-level enrollment data
}
\keyword{internal}
