% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_district_enr}
\alias{process_district_enr}
\title{Process district-level enrollment data}
\usage{
process_district_enr(df, end_year)
}
\arguments{
\item{df}{Raw district data frame}

\item{end_year}{School year end}
}
\value{
Processed district data frame
}
\description{
Process district-level enrollment data
}
\keyword{internal}
