% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_staar.R
\name{read_cache_staar}
\alias{read_cache_staar}
\title{Read STAAR cache}
\usage{
read_cache_staar(year, level, grade, subject, cache_type)
}
\arguments{
\item{year}{Assessment year}

\item{level}{"district" or "campus"}

\item{grade}{Grade level}

\item{subject}{"english" or "spanish"}

\item{cache_type}{"tidy" or "processed"}
}
\value{
Cached data frame or NULL if cache doesn't exist
}
\description{
Read STAAR cache
}
\keyword{internal}
