% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_staar.R
\name{staar_available_years}
\alias{staar_available_years}
\title{Get available STAAR years}
\usage{
staar_available_years()
}
\value{
Integer vector of available years
}
\description{
Returns a vector of years for which STAAR data is available.
}
\keyword{internal}
