% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_graduation.R
\name{tidy_grad}
\alias{tidy_grad}
\title{Tidy graduation rate data}
\usage{
tidy_grad(df)
}
\arguments{
\item{df}{A wide data.frame of processed graduation rate data}
}
\value{
A long data.frame of tidied graduation rate data
}
\description{
Transforms wide graduation rate data to long format with subgroup column.
}
\examples{
\dontrun{
wide_data <- fetch_grad(2024, tidy = FALSE)
tidy_data <- tidy_grad(wide_data)
}
}
