% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_staar.R
\name{write_cache_staar}
\alias{write_cache_staar}
\title{Write STAAR cache}
\usage{
write_cache_staar(data, year, level, grade, subject, cache_type)
}
\arguments{
\item{data}{Data frame to cache}

\item{year}{Assessment year}

\item{level}{"district" or "campus"}

\item{grade}{Grade level}

\item{subject}{"english" or "spanish"}

\item{cache_type}{"tidy" or "processed"}
}
\description{
Write STAAR cache
}
\keyword{internal}
